Moved all backends from backend packages to respected view and modal popup files.
