+++ sclient/src/main/scala/synereo/client/modalpopups/ConfirmIntroReqModal.scala
@@ -1,24 +1,24 @@
 package synereo.client.modalpopups
 
-import japgolly.scalajs.react._
 import japgolly.scalajs.react.extra.OnUnmount
 import japgolly.scalajs.react.vdom.prefix_<^._
 import synereo.client.components.Bootstrap.{Button, CommonStyle, _}
 import synereo.client.components.{GlobalStyles, _}
-import japgolly.scalajs.react
-import shared.dtos.{IntroConfirmReq, Introduction}
+import shared.dtos.{Introduction}
 import diode.AnyAction._
-
 import scalacss.Defaults._
 import scalacss.ScalaCssReact._
 import scala.language.reflectiveCalls
 import synereo.client.css.{DashboardCSS, NewMessageCSS}
-import synereo.client.handlers.UpdateIntroductionsModel
-import synereo.client.modalpopupbackends.ConfirmIntroReqBackend
-import synereo.client.services.SYNEREOCircuit
-
-import scala.concurrent.ExecutionContext.Implicits.global
 import scala.scalajs.js.JSON
+import japgolly.scalajs.react
+import japgolly.scalajs.react.{Callback, _}
+import shared.dtos.IntroConfirmReq
+import synereo.client.components._
+import synereo.client.components.Bootstrap._
+import synereo.client.handlers.UpdateIntroductionsModel
+import synereo.client.services.SYNEREOCircuit
+import diode.AnyAction._
 
 /**
   * Created by mandar.k on 6/29/2016.
@@ -80,6 +80,42 @@ object ConfirmIntroReqForm {
 
   case class State(confirmIntroReq: Boolean = false)
 
+  //    toDo: Think of some better logic to reduce verbosity in accept on form submit or reject --> hide like get  event target source and modify only accepted field of case class
+  case class ConfirmIntroReqBackend(t: BackendScope[Props, State]) {
+    def hide: Callback = Callback {
+      //      val uri = SYNEREOCircuit.zoom(_.sessionRootModel.sessionUri).value
+      val uri = SYNEREOCircuit.zoom(_.sessionRootModel.sessionUri).value
+      val props = t.props.runNow()
+      val introConfirmReq = IntroConfirmReq(uri, alias = "alias", props.introduction.introSessionId, props.introduction.correlationId, accepted = false)
+      //      CoreApi.postIntroduction(introConfirmReq).onComplete {
+      //        case Success(response) =>
+      //
+      //      }
+      SYNEREOCircuit.dispatch(UpdateIntroductionsModel(introConfirmReq))
+      jQuery(t.getDOMNode()).modal("hide")
+    }
+
+    def hideModal(): Unit = {
+      jQuery(t.getDOMNode()).modal("hide")
+    }
+
+    def mounted(props: ConfirmIntroReqForm.Props): Callback = Callback {
+    }
+
+    def submitForm(e: ReactEventI): react.Callback = {
+      e.preventDefault()
+      //      val uri = SYNEREOCircuit.zoom(_.sessionRootModel.sessionUri).value
+      val uri = SYNEREOCircuit.zoom(_.sessionRootModel.sessionUri).value
+      val props = t.props.runNow()
+      val content = IntroConfirmReq(uri, alias = "alias", props.introduction.introSessionId, props.introduction.correlationId, accepted = true)
+      SYNEREOCircuit.dispatch(UpdateIntroductionsModel(content))
+      t.modState(s => s.copy(confirmIntroReq = true))
+    }
+
+    def formClosed(state: ConfirmIntroReqForm.State, props: ConfirmIntroReqForm.Props): Callback = {
+      props.submitHandler(/*state.postMessage*/)
+    }
+  }
 
 
   private val component = ReactComponentB[Props]("PostNewMessage")
