+++ sclient/src/main/scala/synereo/client/modalpopups/LoginFailed.scala
@@ -5,17 +5,16 @@ package synereo.client.modalpopups
   */
 
 import synereo.client.components.{GlobalStyles, Icon}
-import japgolly.scalajs.react._
 import japgolly.scalajs.react.vdom.prefix_<^._
 import synereo.client.components.Bootstrap.Modal
-import synereo.client.components._
 import synereo.client.css.{LoginCSS, SignupCSS, SynereoCommanStylesCSS}
-
 import scala.util.{Failure, Success}
 import scalacss.ScalaCssReact._
 import scala.language.reflectiveCalls
+import japgolly.scalajs.react._
+import synereo.client.components._
 import synereo.client.components.Bootstrap._
-import synereo.client.modalpopupbackends.LoginFailedBackend
+
 
 object LoginFailed {
   // shorthand fo
@@ -24,6 +23,19 @@ object LoginFailed {
   case class Props(submitHandler: () => Callback, loginErrorMessage: String = "")
 
   case class State()
+
+  class LoginFailedBackend(t: BackendScope[Props, State]) {
+
+    def hide = Callback {
+      jQuery(t.getDOMNode()).modal("hide")
+    }
+
+    def formClosed(state: LoginFailed.State, props: LoginFailed.Props): Callback = {
+      // call parent handler with the new item and whether form was OK or cancelled
+      props.submitHandler()
+    }
+
+  }
 
   private val component = ReactComponentB[Props]("LoginFailed")
     .initialState_P(p => State())
