+++ sclient/src/main/scala/synereo/client/modalpopups/VerifyTokenModal.scala
@@ -8,7 +8,10 @@ import synereo.client.components.Bootstrap.Modal
 import synereo.client.css.{SignupCSS}
 import scalacss.ScalaCssReact._
 import scala.language.reflectiveCalls
-import synereo.client.modalpopupbackends.{VerifyTokenModalBackend => Backend}
+import japgolly.scalajs.react._
+import synereo.client.components._
+import scala.language.reflectiveCalls
+import synereo.client.components.Bootstrap._
 
 /**
   * Created by Mandar on 4/19/2016.
@@ -20,9 +23,37 @@ object VerifyTokenModal {
 
   case class State(emailValidationModel: EmailValidationModel, accountValidationFailed: Boolean = false, showLoginForm: Boolean = false)
 
+  class VerifyTokenModalBackend(t: BackendScope[Props, State]) {
+
+    def submitForm(e: ReactEventI) = {
+      e.preventDefault()
+      // mark it as NOT cancelled (which is the default)
+      t.modState(s => s.copy(accountValidationFailed = true))
+    }
+
+    def hideModal = {
+      // instruct Bootstrap to hide the modal
+      jQuery(t.getDOMNode()).modal("hide")
+      t.modState(s => s.copy(showLoginForm = true))
+    }
+
+    def updateToken(e: ReactEventI) = {
+      // update TodoItem content
+      val value = e.target.value
+      t.modState(s => s.copy(emailValidationModel = s.emailValidationModel.copy(token = value)))
+    }
+
+
+
+    def formClosed(state: VerifyTokenModal.State, props: VerifyTokenModal.Props): Callback = {
+      // call parent handler with the new item and whether form was OK or cancelled
+      props.submitHandler(state.emailValidationModel, state.accountValidationFailed,state.showLoginForm)
+    }
+  }
+
   private val component = ReactComponentB[Props]("ConfirmAccountCreation")
     .initialState_P(p => State(new EmailValidationModel("")))
-    .backend(new Backend(_))
+    .backend(new VerifyTokenModalBackend(_))
     .renderPS((t, P, S) => {
       val headerText = "Verify Token"
       Modal(
